feat: rename genomic axis
