+++ tsconfig.json
@@ -28,7 +28,7 @@
       "@gosling-lang/higlass-schema": ["src/higlass-schema/index.ts"],
       "@gosling-lang/gosling-theme": ["src/gosling-theme/index.ts"],
       "@gosling-lang/gosling-track": ["./src/tracks/gosling-track/index.ts"],
-      "@gosling-lang/gosling-genomic-axis": ["./src/tracks/gosling-genomic-axis/index.ts"],
+      "@gosling-lang/genomic-axis": ["src/tracks/genomic-axis/index.ts"],
       "@gosling-lang/brush-circular": ["src/tracks/brush-circular/index.ts"],
       "@gosling-lang/brush-linear": ["src/tracks/brush-linear/index.ts"],
       "@gosling-lang/dummy-track": ["./src/tracks/dummy-track/index.ts"],
